Add files via upload
This is my contribution to our final output in experimental psychology, in here I was assigned the data analysis portion, the data visualization dashboard, the clean data sheet, and the results and discussions section of our IMRAD formatted research paper. 

The following are the four files containing my portion of work in the project.  First of all is the data analysis portion titled experimental data analysis which was analyzed in Microsoft Excel courtesy of the independent t test in the data analysis toolpack add-in, equipped with a data visualization dashboard  that was created via Pivot Table and Pivot chart. The next file is the clean data sheet that was cleaned and analyzed in Microsoft Excel, the spreadsheet consists of four columns, the experimental group and it's participants along with the control group and it's participants. The third file consists of the reports and discussion section which I wrote in order to explain the data analysis portion and data visualization dashboard thoughtfully.  The final file is the finished research paper written in a summarized IMRAD format courtesy of our university, the results and discussion section equipped with the data analysis and dashboard can be found at the discussion section the second to last part before the recommendations of the paper.